+++ pom.xml
@@ -115,8 +115,9 @@
                     <execution>
                         <goals>
                             <goal>compile</goal>
-                            <goal>testCompile</goal>
+<!--                            <goal>testCompile</goal>-->
                         </goals>
+
                         <configuration>
                             <recompileMode>incremental</recompileMode>
                             <args>
