+++ src/main/resources/remote.conf
@@ -1,11 +1,11 @@
 akka {
   remote {
     netty.tcp {
-      hostname = "174.63.26.213"
-      port = 12382
+//      hostname = "174.63.26.213"
+//      port = 12382
 
-      bind-hostname = "10.0.0.34"
-      bind-port = 3451
+      hostname = "10.0.0.34"
+      port = 12382
     }
   }
 }
