+++ src/main/resources/remoting_example.conf
@@ -16,11 +16,11 @@ akka {
     transport = "akka.remote.netty.NettyRemoteTransport"
 
     netty.tcp {
-      hostname = "174.63.26.213"
-      port = 12381
+//      hostname = "174.63.26.213"
+//      port = 12381
 
-      bind-hostname = "10.0.0.34"
-      bind-port = 65534
+      hostname = "10.0.0.34"
+      port = 12381
     }
 
     log-sent-messages = on
@@ -31,6 +31,8 @@ akka {
 
 nodes {
   locations = [
-    "akka.tcp://RemoteEvvoNode@174.63.26.213:3451"
+//    "akka.tcp://RemoteEvvoNode@35.245.166.99:12382"
+    "akka.tcp://RemoteEvvoNode@10.0.0.34:12382"
+//    "akka.tcp://RemoteEvvoNode@174.63.26.213:12382"
   ]
 }
